fix : 파이어베이스 자료형에 맞게 뷰모델 메서드 내부 cast 수정

+++ hifi_Shopping/app/src/main/java/com/hifi/hifi_shopping/user/vm/PointViewModel.kt
@@ -19,7 +19,7 @@ class PointViewModel  : ViewModel() {
         PointRepository.getPointListByUserEmail(userIdx) {
             for(c1 in it.result.children) {
                 val userIdx = c1.child("userIdx").value as String
-                val amount = (c1.child("amount").value as String).toLong()
+                val amount = c1.child("amount").value as String
                 val date = c1.child("date").value as String
                 val context = c1.child("context").value as String
